Reword Arch Linux Install Instructions
Problems I see:
 - Yaourt is unmaintained and shouldn't be used anymore.
 - It is recommended that new arch users build AUR packages from scratch instead of using helpers. They need to understand what is going on behind the scenes before using a helper.
 - The AUR can potentially hold malicious code and build scripts so we shouldn't encourage users to blindly install things from it.

My Solutions:
Removed all mention to an AUR helper. People who are comfortable with the AUR only need the package name to install it.  For those who are inexperienced, links to the wiki have been given. They need to understand exactly what the AUR is and how to use it safely before installing packages.